// frontend/src/generated-client/api.ts
export interface Feature {
    /**
     * 
     * @type {number}
     * @memberof Feature
     */
    'featureID'?: number;
    /**
     * 
     * @type {string}
     * @memberof Feature
     */
    'featureName': string;
    /**
     * 
     * @type {number}
     * @memberof Feature
     */
    'companyID': number;
}
// ...
export class FeaturesApi extends BaseAPI {
    /**
     * 
     * @param {number} companyId 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof FeaturesApi
     */
    public featuresCompanyCompanyIdGet(companyId: number, options?: AxiosRequestConfig) {
        return FeaturesApiFp(this.configuration).featuresCompanyCompanyIdGet(companyId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} featureId 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof FeaturesApi
     */
    public featuresFeatureIdDelete(featureId: number, options?: AxiosRequestConfig) {
        return FeaturesApiFp(this.configuration).featuresFeatureIdDelete(featureId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof FeaturesApi
     */
    public featuresGet(options?: AxiosRequestConfig) {
        return FeaturesApiFp(this.configuration).featuresGet(options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {Feature} [feature] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof FeaturesApi
     */
    public featuresPost(feature?: Feature, options?: AxiosRequestConfig) {
        return FeaturesApiFp(this.configuration).featuresPost(feature, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {Feature} [feature] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof FeaturesApi
     */
    public featuresPut(feature?: Feature, options?: AxiosRequestConfig) {
        return FeaturesApiFp(this.configuration).featuresPut(feature, options).then((request) => request(this.axios, this.basePath));
    }
}